requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
feedparser==6.0.12
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
sgmllib3k==1.0.0
urllib3==2.6.3
